Etain Final Project Code/requirements.txt
mysql-connector-python==8.0.18
Flask==1.1.1